Make Java binding for cli util compile
