+++ src/mlpack/bindings/java/cli_util.cpp
@@ -1,27 +1,11 @@
-#ifndef MLPACK_BINDINGS_JAVA_MLPACK_CLI_UTIL_HPP
-#define MLPACK_BINDINGS_JAVA_MLPACK_CLI_UTIL_HPP
-
-#include <mlpack/core/util/cli.hpp>
+#include "cli_util.hpp"
 
 namespace mlpack {
 namespace util {
 
-void DeleteArray(const void* p) 
+void DeleteArray(void* p) 
 {
-  delete[] static_cast<const char*>(p);
-}
-
-template <typename T>
-void SetMatParam(const char* name, const T* data, size_t rows, size_t columns)
-{
-  arma::Mat<T> m(data, rows, columns, false, true);
-  CLI::GetParam<decltype(m)>(name) = std::move(m);
-}
-
-template <typename T>
-void SetParam(const char* name, T value) 
-{
-  CLI::GetParam<T>(name) = std::move(value);
+  delete[] static_cast<unsigned char*>(p);
 }
 
 double* GetMatParamData(const char* name) 
@@ -56,5 +40,3 @@ size_t GetMatParamLength(const char* name)
 
 }
 }
-
-#endif
