+++ src/mlpack/methods/pca/CMakeLists.txt
@@ -20,4 +20,5 @@ set(MLPACK_SRCS ${MLPACK_SRCS} ${DIR_SRCS} PARENT_SCOPE)
 
 add_cli_executable(pca)
 add_python_binding(pca)
+add_java_binding(pca PCA)
 add_markdown_docs(pca "cli;python" "transformations")
